backend-2023-main/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
boto3==1.28.27
botocore==1.31.27
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
deprecation==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
img2pdf==0.4.4
jmespath==1.0.1
kombu==5.3.1
lxml==4.9.3
packaging==23.2
pdfkit==1.0.0
pikepdf==8.5.0
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycryptodome==3.19.0
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
qrcode==7.4.2
redis==5.0.0
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
